requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
gspread==5.0.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
python-dotenv==0.19.1
openai>=1.0.0
uuid>=1.30